re-enable react player test. fail player state from error controller if no transition available.
