+++ core/player/src/controllers/error/utils/index.ts
@@ -1,0 +1,1 @@
+export * from "./isErrorWithMetadata";
