+++ react/player/src/__tests__/app.test.tsx
@@ -8,7 +8,7 @@ import {
   configure,
 } from "@testing-library/react";
 import type { InProgressState } from "@player-ui/player";
-// import { makeFlow } from "@player-ui/make-flow";
+import { makeFlow } from "@player-ui/make-flow";
 import { ReactPlayer } from "..";
 import { simpleFlow, SimpleAssetPlugin } from "./helpers/simple-asset-plugin";
 
@@ -46,15 +46,14 @@ describe("ReactPlayer React", () => {
     expect(getNodeText(viewNode!)).toBe("Initial Value");
   });
 
-  // TODO: Re-enable test when error handling work is complete
-  // test("fails flow when UI throws error", async () => {
-  //   const rp = new ReactPlayer();
-  //   const response = rp.start(makeFlow({ type: "err", id: "Error" }));
-  //   act(() => {
-  //     render(<rp.Component />);
-  //   });
-  //   await expect(response).rejects.toThrow();
-  // });
+  test("fails flow when UI throws error", async () => {
+    const rp = new ReactPlayer();
+    const response = rp.start(makeFlow({ type: "err", id: "Error" }));
+    act(() => {
+      render(<rp.Component />);
+    });
+    await expect(response).rejects.toThrow();
+  });
 
   test("updates the react comp when view updates", async () => {
     const rp = new ReactPlayer({
