+++ react/player/src/asset/index.tsx
@@ -124,115 +124,3 @@ export const ReactAsset = (
     </ErrorBoundary>
   );
 };
-
-type AssetClassState = { currentError: Error };
-export class ReactAssetClass extends React.Component<
-  AssetType<string> | AssetWrapper<AssetType<string>>,
-  AssetClassState
-> {
-  static contextType: typeof AssetContext = AssetContext;
-  declare context: React.ContextType<typeof AssetContext>;
-
-  static getDerivedStateFromError(err: Error): AssetClassState {
-    return { currentError: err };
-  }
-
-  override componentDidCatch(error: Error, errorInfo: React.ErrorInfo): void {
-    // const unwrapped = this.getUnwrappedAssetFromProps();
-    // if (error instanceof AssetRenderError) {
-    //   error.addAssetParent(unwrapped);
-    //   throw error;
-    // } else {
-    //   throw new AssetRenderError(unwrapped, "Failed to render asset", error);
-    // }
-  }
-
-  private getUnwrappedAssetFromProps(): AssetType<string> {
-    const props: AssetType<string> | AssetWrapper<AssetType<string>> =
-      this.props;
-
-    if (isAssetUnwrapped(props)) {
-      return props;
-    } else if ("asset" in props) {
-      return props.asset;
-    }
-
-    throw Error(
-      `Cannot determine asset type for props: ${JSON.stringify(props)}`,
-    );
-  }
-
-  render(): React.ReactNode {
-    const props = this.props;
-    const { registry } = this.context;
-
-    const unwrapped = this.getUnwrappedAssetFromProps();
-
-    if (typeof unwrapped !== "object") {
-      throw Error(
-        `Asset was not an object got (${typeof unwrapped}) instead: ${unwrapped}`,
-      );
-    }
-
-    if (unwrapped.type === undefined) {
-      const info =
-        unwrapped.id === undefined
-          ? JSON.stringify(props)
-          : `id: ${unwrapped.id}`;
-      throw Error(`Asset is missing type for ${info}`);
-    }
-
-    if (!registry || registry.isRegistryEmpty()) {
-      throw Error(`No asset found in registry. This could happen for one of the following reasons: \n
-      1. You might have no assets registered or no plugins added to the Player instance. \n
-      2. You might have mismatching versions of React Asset Registry Context. \n
-      See https://player-ui.github.io/latest/tools/cli#player-dependency-versions-check for tips about how to debug and fix this problem`);
-    }
-
-    const Impl = registry?.get(unwrapped);
-
-    if (!Impl) {
-      const matchList: object[] = [];
-
-      registry.forEach((asset) => {
-        matchList.push(asset.key);
-      });
-
-      const typeList = matchList.map(
-        (match) => JSON.parse(JSON.stringify(match)).type,
-      );
-
-      const similarType = typeList.reduce((prev, curr) => {
-        const next = {
-          value: leven(unwrapped.type, curr),
-          type: curr,
-        };
-
-        if (prev !== undefined && prev.value < next.value) {
-          return prev;
-        }
-
-        return next;
-      }, undefined);
-
-      throw Error(
-        `No implementation found for id: ${unwrapped.id} type: ${unwrapped.type}. Did you mean ${similarType.type}? \n 
-      Registered Asset matching functions are listed below: \n
-      ${JSON.stringify(matchList)}`,
-      );
-    }
-
-    const error = this.state?.currentError;
-    if (error) {
-      this.setState({});
-      if (error instanceof AssetRenderError) {
-        error.addAssetParent(unwrapped);
-        throw error;
-      } else {
-        throw new AssetRenderError(unwrapped, "Failed to render asset", error);
-      }
-    }
-
-    return <Impl key={unwrapped.id} {...unwrapped} />;
-  }
-}
